BorderBoxControl: Refactor storybook to TypeScript
